requirements.txt
aiokafka==0.8.1
alembic==1.12.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
asyncpg==0.29.0
certifi==2023.7.22
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
exceptiongroup==1.1.3
fastapi==0.104.0
graphql-core==3.2.3
greenlet==3.0.0
h11==0.14.0
httpcore==1.0.1
httpx==0.25.1
idna==3.4
kafka-python==2.0.2
Mako==1.3.0
MarkupSafe==2.1.3
motor==3.3.1
packaging==23.2
pydantic==2.4.2
pydantic_core==2.10.1
pymongo==4.6.0
python-dateutil==2.8.2
python-multipart==0.0.6
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.22
starlette==0.27.0
strawberry-graphql==0.215.1
typing_extensions==4.8.0
uvicorn==0.23.2
